+++ Sources/Servers/Project/ProjectMCPServer.swift
@@ -33,6 +33,7 @@ public enum ProjectToolName: String, CaseIterable, Sendable {
     case addTestPlanToScheme = "add_test_plan_to_scheme"
     case removeTestPlanFromScheme = "remove_test_plan_from_scheme"
     case listTestPlans = "list_test_plans"
+    case searchTestPlans = "search_test_plans"
     case setTestTargetApplication = "set_test_target_application"
     case renameGroup = "rename_group"
     case addDependency = "add_dependency"
@@ -165,6 +166,7 @@ public struct ProjectMCPServer: Sendable {
         let addTestPlanToSchemeTool = AddTestPlanToSchemeTool(pathUtility: pathUtility)
         let removeTestPlanFromSchemeTool = RemoveTestPlanFromSchemeTool(pathUtility: pathUtility)
         let listTestPlansTool = ListTestPlansTool(pathUtility: pathUtility)
+        let searchTestPlansTool = SearchTestPlansTool(pathUtility: pathUtility)
         let setTestTargetApplicationTool = SetTestTargetApplicationTool(pathUtility: pathUtility)
         let renameGroupTool = RenameGroupTool(pathUtility: pathUtility)
         let addDependencyTool = AddDependencyTool(pathUtility: pathUtility)
@@ -255,6 +257,7 @@ public struct ProjectMCPServer: Sendable {
                 addTestPlanToSchemeTool.tool(),
                 removeTestPlanFromSchemeTool.tool(),
                 listTestPlansTool.tool(),
+                searchTestPlansTool.tool(),
                 setTestTargetApplicationTool.tool(),
                 renameGroupTool.tool(),
                 addDependencyTool.tool(),
@@ -372,6 +375,8 @@ public struct ProjectMCPServer: Sendable {
                     return try removeTestPlanFromSchemeTool.execute(arguments: arguments)
                 case .listTestPlans:
                     return try listTestPlansTool.execute(arguments: arguments)
+                case .searchTestPlans:
+                    return try searchTestPlansTool.execute(arguments: arguments)
                 case .setTestTargetApplication:
                     return try setTestTargetApplicationTool.execute(arguments: arguments)
                 case .renameGroup: